fix[database]: rename sql files to order them
when the files of the `sql` directory are indexed by `init_db.py`
scipt the order they are executed is wrong. Since some tables depend
on the existence of others they have to be executed in a particular
order. Therefore files were prefixed with the index of the correct
order.